pyproject.toml
[project]
name = "elasticsearch-quest"
version = "0.0.1"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "elasticsearch>=8.17.2",
    "gradio>=5.25.0",
    "openai-agents>=0.0.12",
    "python-dotenv>=1.1.0",
]

[dependency-groups]
dev = [
    "pytest-asyncio>=0.26.0",
    "pytest>=8.3.5",
    "ruff>=0.11.5",
]

[tool.ruff]
line-length = 88

[tool.ruff.lint]
extend-select = ["E", "W", "F", "I"]
ignore = []

[tool.ruff.lint.isort]
combine-as-imports = true
known-first-party = []
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]
split-on-trailing-comma = true
